The Task
You will answer a series of questions using T-SQL. These questions are based on a database called WideWorldImporters, which must be downloaded and installed locally on your computer.

A total of 25 questions need to be answered, with each question carrying a point value of 1, 2, or 3 points. The maximum possible score for all questions combined is 50 points.